Added ability to view and add members to related projects
