+++ components/ProjectDetails/MainContentPD/ProjectDetailsInside/styles.module.css
@@ -1,0 +1,17 @@
+@import url('https://fonts.googleapis.com/css2?family=Train+One&display=swap');
+
+.container {
+    /* border: 1px solid rgb(255, 0, 0); */
+    background-color: white;
+    height: auto !important;
+    z-index: -10 !important;
+}
+
+.header {
+    position: sticky !important;
+    /* border: 1px solid red; */
+    top: 5.3%;
+    margin-left: 2px;
+    background-color: white;
+    z-index: 0.1 !important;
+}
